+++ Cargo.toml
@@ -4,7 +4,6 @@ members = [
   "tiledb/api",
   "tiledb/common",
   "tiledb/proc-macro",
-  "tiledb/proptest-integrations",
   "tiledb/queries",
   "tiledb/serde",
   "tiledb/sys",
@@ -20,7 +19,6 @@ default-members = [
   "tiledb/api",
   "tiledb/common",
   "tiledb/proc-macro",
-  "tiledb/proptest-integrations",
   "tiledb/queries",
   "tiledb/serde",
   "tiledb/sys",
@@ -37,6 +35,7 @@ anyhow = "1.0"
 arrow = { version = "52.0.0", features = ["prettyprint"] }
 arrow-schema = { version = "52.0.0" }
 cells = { path = "test-utils/cells", version = "0.1.0" }
+itertools = "0"
 num-traits = "0.2"
 paste = "1.0"
 proptest = { version = "1.0.0" }
