+++ tiledb/api/Cargo.toml
@@ -10,7 +10,7 @@ path = "src/lib.rs"
 [dependencies]
 anyhow = { workspace = true }
 arrow = { workspace = true, optional = true }
-itertools = "0"
+itertools = { workspace = true }
 num-traits = { workspace = true, optional = true }
 paste = { workspace = true }
 proptest = { workspace = true, optional = true }
