+++ tiledb/common/src/datatype/mod.rs
@@ -149,7 +149,7 @@ pub enum Datatype {
     GeometryWkt,
 }
 
-const DATATYPES: [Datatype; 43] = [
+pub const DATATYPES: [Datatype; 43] = [
     Datatype::Int32,
     Datatype::Int64,
     Datatype::Float32,
@@ -193,6 +193,75 @@ const DATATYPES: [Datatype; 43] = [
     Datatype::Boolean,
     Datatype::GeometryWkb,
     Datatype::GeometryWkt,
+];
+
+pub const DENSE_DIMENSION_DATATYPES: [Datatype; 30] = [
+    Datatype::Int8,
+    Datatype::Int16,
+    Datatype::Int32,
+    Datatype::Int64,
+    Datatype::UInt8,
+    Datatype::UInt16,
+    Datatype::UInt32,
+    Datatype::UInt64,
+    Datatype::DateTimeYear,
+    Datatype::DateTimeMonth,
+    Datatype::DateTimeWeek,
+    Datatype::DateTimeDay,
+    Datatype::DateTimeHour,
+    Datatype::DateTimeMinute,
+    Datatype::DateTimeSecond,
+    Datatype::DateTimeMillisecond,
+    Datatype::DateTimeMicrosecond,
+    Datatype::DateTimeNanosecond,
+    Datatype::DateTimePicosecond,
+    Datatype::DateTimeFemtosecond,
+    Datatype::DateTimeAttosecond,
+    Datatype::TimeHour,
+    Datatype::TimeMinute,
+    Datatype::TimeSecond,
+    Datatype::TimeMillisecond,
+    Datatype::TimeMicrosecond,
+    Datatype::TimeNanosecond,
+    Datatype::TimePicosecond,
+    Datatype::TimeFemtosecond,
+    Datatype::TimeAttosecond,
+];
+
+pub const SPARSE_DIMENSION_DATATYPES: [Datatype; 33] = [
+    Datatype::Int8,
+    Datatype::Int16,
+    Datatype::Int32,
+    Datatype::Int64,
+    Datatype::UInt8,
+    Datatype::UInt16,
+    Datatype::UInt32,
+    Datatype::UInt64,
+    Datatype::Float32,
+    Datatype::Float64,
+    Datatype::DateTimeYear,
+    Datatype::DateTimeMonth,
+    Datatype::DateTimeWeek,
+    Datatype::DateTimeDay,
+    Datatype::DateTimeHour,
+    Datatype::DateTimeMinute,
+    Datatype::DateTimeSecond,
+    Datatype::DateTimeMillisecond,
+    Datatype::DateTimeMicrosecond,
+    Datatype::DateTimeNanosecond,
+    Datatype::DateTimePicosecond,
+    Datatype::DateTimeFemtosecond,
+    Datatype::DateTimeAttosecond,
+    Datatype::TimeHour,
+    Datatype::TimeMinute,
+    Datatype::TimeSecond,
+    Datatype::TimeMillisecond,
+    Datatype::TimeMicrosecond,
+    Datatype::TimeNanosecond,
+    Datatype::TimePicosecond,
+    Datatype::TimeFemtosecond,
+    Datatype::TimeAttosecond,
+    Datatype::StringAscii,
 ];
 
 impl Datatype {
