--- tiledb/proptest-integrations/src/array/attribute.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-mod tests {}
