--- tiledb/proptest-integrations/src/array/mod.rs
@@ -1,1 +1,0 @@
-mod attribute;
