--- tiledb/proptest-integrations/src/lib.rs
@@ -1,1 +1,0 @@
-mod array;
