+++ tiledb/serde/Cargo.toml
@@ -5,6 +5,7 @@ rust-version.workspace = true
 version.workspace = true
 
 [dependencies]
+itertools = { workspace = true }
 num-traits = { workspace = true, optional = true }
 proptest = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
